requirements.txt
schedule
requests
python-dotenv
plyer
